# docs/api/swagger.yaml
basePath: /
definitions:
...
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.Code:
    enum:
    - INVALID_PAYLOAD
    - VALIDATION_FAILED
    - UNAUTHENTICATED
    - FORBIDDEN
    - RESOURCE_NOT_FOUND
    - CONFLICT
    - RATE_LIMITED
    - PAYLOAD_TOO_LARGE
    - SEARCH_UNAVAILABLE
    - INTERNAL_ERROR
    - NOT_IMPLEMENTED
    type: string
    x-enum-varnames:
    - CodeInvalidPayload
    - CodeValidationFailed
    - CodeUnauthenticated
    - CodeForbidden
    - CodeNotFound
    - CodeConflict
    - CodeRateLimited
    - CodePayloadTooLarge
    - CodeSearchUnavailable
    - CodeInternal
    - CodeNotImplemented
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.InvalidParam:
    properties:
      code:
        type: string
      field:
        type: string
    type: object
...
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails:
    properties:
      code:
        $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.Code'
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.InvalidParam'
        type: array
      instance:
        type: string
      request_id:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse:
    properties:
      data: {}
      request_id:
        type: string
    type: object
...
  internal_support_delivery_http.createTicketRequest:
    properties:
      body:
        type: string
      email:
        type: string
      subject:
        type: string
    type: object
  internal_support_delivery_http.messageRequest:
    properties:
      body:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for AquaWheel e-commerce project.
  title: AquaWheel Store API
  version: "1.0"
paths:
...
  /api/v1/support/tickets:
    post:
      consumes:
      - application/json
      description: Public intake, protected by a distributed per-IP and per-email
        quota. A guest supplies an email; for an authenticated customer it is resolved
        server-side.
      parameters:
      - description: Ticket
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_support_delivery_http.createTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: Open a support ticket
      tags:
      - Support v1
  /api/v1/support/tickets/{id}/messages:
    post:
      consumes:
      - application/json
      description: 'Requires authentication: ticket ownership is checked against the
        authenticated subject, so a guest ticket cannot be read or extended by guessing
        its ID.'
      parameters:
      - description: Ticket UUID
        in: path
        name: id
        required: true
        type: string
      - description: Message
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_support_delivery_http.messageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: Reply to your own support ticket
      tags:
      - Support v1
...
  /api/wishlist:
    get:
      description: Scoped to the authenticated customer, or to the anonymous cart_session
        cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List the current buyer's wishlist
      tags:
      - Wishlist
  /api/wishlist/{variant_id}:
    delete:
      parameters:
      - description: Product variant UUID
        in: path
        name: variant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove a product variant from the wishlist
      tags:
      - Wishlist
    post:
      parameters:
      - description: Product variant UUID
        in: path
        name: variant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a product variant to the wishlist
      tags:
      - Wishlist
securityDefinitions:
  bearerAuth:
    description: 'Введіть токен у форматі: Bearer {your_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...